Split up DirectILLAutoProcessTest
RE #34850
